Refactor `AlgebraicMap`
