+++ symbolic-base/src/ZkFold/Base/Protocol/Protostar/AlgebraicMap.hs
@@ -1,15 +1,10 @@
-{-# LANGUAGE AllowAmbiguousTypes     #-}
-{-# LANGUAGE TypeOperators           #-}
-{-# LANGUAGE UndecidableInstances    #-}
-{-# LANGUAGE UndecidableSuperClasses #-}
-{-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# LANGUAGE TypeOperators #-}
 
-module ZkFold.Base.Protocol.Protostar.ArithmeticCircuit where
-
+module ZkFold.Base.Protocol.Protostar.AlgebraicMap where
 
 import           Data.Functor.Rep                                    (tabulate)
 import           Data.List                                           (foldl')
-import           Data.Map.Strict                                     (Map, elems, fromList)
+import           Data.Map.Strict                                     (Map, fromList)
 import qualified Data.Map.Strict                                     as M
 import           GHC.IsList                                          (toList)
 import           Prelude                                             (fmap, zip, ($), (++), (.), (<$>))
@@ -21,34 +16,30 @@ import qualified ZkFold.Base.Algebra.Polynomials.Multivariate        as PM
 import           ZkFold.Base.Algebra.Polynomials.Multivariate
 import qualified ZkFold.Base.Data.Vector                             as V
 import           ZkFold.Base.Data.Vector                             (Vector)
-import qualified ZkFold.Base.Protocol.Protostar.SpecialSound         as SPS
 import           ZkFold.Symbolic.Class
 import           ZkFold.Symbolic.Compiler
 import           ZkFold.Symbolic.Compiler.ArithmeticCircuit.Internal
 import           ZkFold.Symbolic.Data.Eq
 
-instance (KnownNat n, Arithmetic a) => SPS.SpecialSoundProtocol a (ArithmeticCircuit a (Vector n) o) where
+-- | Algebraic map is a much more versatile and powerful tool when used separatey from SpecialSoundProtocol.
+-- It calculates a system of equations @[f]@ defining @a@ in some way.
+-- If @f@ is a number or a field element, then the result is a vector of polynomial values.
+-- However, @f@ can be a polynomial, in which case the result will be a system of polynomials.
+-- This polymorphism is exploited in the AccumulatorScheme prover.
+--
+class AlgebraicMap f a where
+    type MapInput f a
+    type MapMessage f a
+    type Degree a :: Natural
+    -- ^ d in the paper, the verifier degree
 
-    type Witness a (ArithmeticCircuit a (Vector n) o) = ()
-    type Input a (ArithmeticCircuit a (Vector n) o) = Vector n a
-    type ProverMessage a (ArithmeticCircuit a (Vector n) o) = [a]
-    type VerifierMessage a (ArithmeticCircuit a (Vector n) o) = a
-    type VerifierOutput a (ArithmeticCircuit a (Vector n) o)  = [a]
-    type Degree (ArithmeticCircuit a (Vector n) o) = 2
-
-    -- One round for Plonk
-    rounds = P.const 1
-
-    outputLength ac = P.fromIntegral $ M.size (acSystem ac)
-
-    -- The transcript will be empty at this point, it is a one-round protocol.
-    -- Input is arithmetised. We need to combine its witness with the circuit's witness.
-    --
-    prover ac _ pi _ _ = elems $ witnessGenerator ac pi
-
-    -- | Evaluate the algebraic map on public inputs and prover messages and compare it to a list of zeros
-    --
-    verifier rc i pm ts = SPS.algebraicMap rc i pm ts one
+    -- | the algebraic map V_sps computed by the verifier.
+    algebraicMap :: a
+        -> MapInput f a     -- ^ public input
+        -> [MapMessage f a] -- ^ NARK proof witness (the list of prover messages)
+        -> [f]              -- ^ Verifier random challenges
+        -> f                -- ^ Slack variable for padding
+        -> [f]
 
 instance
   ( KnownNat n
@@ -56,10 +47,11 @@ instance
   , Scale a f
   , MultiplicativeMonoid f
   , AdditiveMonoid f
-  ) => SPS.AlgebraicMap f (ArithmeticCircuit a (Vector n) o) where
+  ) => AlgebraicMap f (ArithmeticCircuit a (Vector n) o) where
 
     type MapInput f (ArithmeticCircuit a (Vector n) o) = Vector n f
     type MapMessage f (ArithmeticCircuit a (Vector n) o) = [f]
+    type Degree (ArithmeticCircuit a (Vector n) o) = 2
 
     -- We can use the polynomial system from the circuit as a base for V_sps.
     --
@@ -75,7 +67,7 @@ instance
             varMap x = M.findWithDefault zero x witness
 
             f_sps :: Vector 3 [PM.Poly a (SysVar (Vector n)) Natural]
-            f_sps = degreeDecomposition @(SPS.Degree (ArithmeticCircuit a (Vector n) o)) $ sys
+            f_sps = degreeDecomposition @(Degree (ArithmeticCircuit a (Vector n) o)) $ sys
 
             f_sps_uni :: Vector 3 [f]
             f_sps_uni = fmap (PM.evalPolynomial PM.evalMonomial varMap) <$> f_sps
