+++ symbolic-base/symbolic-base.cabal
@@ -146,11 +146,10 @@ library
       ZkFold.Base.Protocol.Protostar
       ZkFold.Base.Protocol.Protostar.Accumulator
       ZkFold.Base.Protocol.Protostar.AccumulatorScheme
-      ZkFold.Base.Protocol.Protostar.ArithmeticCircuit
+      ZkFold.Base.Protocol.Protostar.AlgebraicMap
       ZkFold.Base.Protocol.Protostar.CommitOpen
       ZkFold.Base.Protocol.Protostar.Commit
       ZkFold.Base.Protocol.Protostar.FiatShamir
-      -- ZkFold.Base.Protocol.Protostar.Fold
       ZkFold.Base.Protocol.Protostar.Function
       ZkFold.Base.Protocol.Protostar.Gate
       ZkFold.Base.Protocol.Protostar.IVC
